+++ local/Cargo.toml
@@ -14,7 +14,8 @@ homepage = "https://pliantdb.dev/"
 [lib]
 
 [features]
-default = ["cli"]
+default = ["full"]
+full = ["cli", "pubsub", "keyvalue"]
 cli = ["structopt"]
 internal-apis = []
 pubsub = ["pliantdb-core/pubsub"]
