+++ local/src/error.rs
@@ -36,7 +36,7 @@ pub enum Error {
 
 impl From<Error> for pliantdb_core::Error {
     fn from(err: Error) -> Self {
-        Self::Storage(err.to_string())
+        Self::Database(err.to_string())
     }
 }
 
@@ -58,7 +58,7 @@ where
     type Output = T;
 
     fn map_err_to_core(self) -> Result<Self::Output, pliantdb_core::Error> {
-        self.map_err(|err| pliantdb_core::Error::Storage(err.into().to_string()))
+        self.map_err(|err| pliantdb_core::Error::Database(err.into().to_string()))
     }
 }
 
@@ -68,7 +68,7 @@ fn test_converting_error() {
     let err: pliantdb_core::Error =
         Error::Serialization(serde_cbor::Error::custom("mymessage")).into();
     match err {
-        pliantdb_core::Error::Storage(storage_error) => {
+        pliantdb_core::Error::Database(storage_error) => {
             assert!(storage_error.contains("mymessage"))
         }
         _ => unreachable!(),
