+++ local/src/open_trees.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 use pliantdb_core::schema::{CollectionName, Schematic};
 
-use crate::{storage::document_tree_name, views::view_invalidated_docs_tree_name, Error};
+use crate::{database::document_tree_name, views::view_invalidated_docs_tree_name, Error};
 
 #[derive(Default)]
 pub struct OpenTrees {
@@ -26,17 +26,18 @@ impl OpenTrees {
     pub fn open_trees_for_document_change(
         &mut self,
         sled: &sled::Db,
+        database: &str,
         collection: &CollectionName,
         schema: &Schematic,
     ) -> Result<(), Error> {
-        self.open_tree(sled, &document_tree_name(collection))?;
+        self.open_tree(sled, &document_tree_name(database, collection))?;
 
         if let Some(views) = schema.views_in_collection(collection) {
             for view in views {
                 let view_name = view.view_name()?;
                 self.open_tree(
                     sled,
-                    &view_invalidated_docs_tree_name(collection, &view_name),
+                    &view_invalidated_docs_tree_name(database, collection, &view_name),
                 )?;
             }
         }
