+++ local/src/views.rs
@@ -21,34 +21,53 @@ pub struct EntryMapping {
 pub mod integrity_scanner;
 pub mod mapper;
 
-pub fn view_entries_tree_name(collection: &CollectionName, view_name: &impl Display) -> String {
-    format!("{}::{}", collection, view_name)
+pub fn view_entries_tree_name(
+    database: &str,
+    collection: &CollectionName,
+    view_name: &impl Display,
+) -> String {
+    format!(
+        "{}::collection::{}::view::{}",
+        database, collection, view_name
+    )
 }
 
 /// Used to store Document ID -> Key mappings, so that when a document is updated, we can remove the old entry.
 pub fn view_document_map_tree_name(
+    database: &str,
     collection: &CollectionName,
     view_name: &impl Display,
 ) -> String {
-    format!("{}::{}::document-map", collection, view_name)
+    format!(
+        "{}::collection::{}::view::{}::document-map",
+        database, collection, view_name
+    )
 }
 
 pub fn view_invalidated_docs_tree_name(
+    database: &str,
     collection: &CollectionName,
     view_name: &impl Display,
 ) -> String {
-    format!("{}::{}::invalidated", collection, view_name)
+    format!(
+        "{}::collection::{}::view::{}::invalidated",
+        database, collection, view_name
+    )
 }
 
 pub fn view_omitted_docs_tree_name(
+    database: &str,
     collection: &CollectionName,
     view_name: &impl Display,
 ) -> String {
-    format!("{}::{}::omitted", collection, view_name)
+    format!(
+        "{}::collection::{}::view::{}::::omitted",
+        database, collection, view_name
+    )
 }
 
-pub fn view_versions_tree_name(collection: &CollectionName) -> String {
-    format!("{}::view-versions", collection)
+pub fn view_versions_tree_name(database: &str, collection: &CollectionName) -> String {
+    format!("{}::collection::{}::view-versions", database, collection)
 }
 
 #[derive(Debug, Clone, Hash, Eq, PartialEq)]
