+++ pliantdb/examples/basic-local.rs
@@ -6,7 +6,7 @@ use pliantdb::{
         schema::{Collection, CollectionName, InvalidNameError, Schematic},
         Error,
     },
-    local::{Configuration, Storage},
+    local::{Configuration, Database},
 };
 use serde::{Deserialize, Serialize};
 
@@ -28,7 +28,7 @@ impl Collection for Message {
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
-    let db = Storage::<Message>::open_local("basic.pliantdb", &Configuration::default()).await?;
+    let db = Database::<Message>::open_local("basic.pliantdb", &Configuration::default()).await?;
     let messages = db.collection::<Message>();
 
     // Insert a new `Message` into the collection. The `push()` method used
